Make the quotation function a constructor
Turns out that we don't need to tie the knot in the syntax to get Lob's
theorem, we can just tie it in the interpreter instead.  Now we don't
need the initial context at all, because we don't need to know how to
quote things until we're interpreting them (at which point, everything
is trivial).